Move records received model with receiver
