+++ app/core/tasks/forward-samples-to-recognition-task.ts
@@ -1,7 +1,7 @@
 import { Task, TaskOutcome, TaskParams } from "nativescript-task-dispatcher/tasks";
 import { DispatchableEvent } from "nativescript-task-dispatcher/events";
 import { getModelType, getSensingDataSource, ModelType, SensingDataSource } from "~/core/mode";
-import { RecordsToProcess } from "~/core/feature-extraction/records-to-process";
+import { RecordsToProcess } from "~/core/receiver/records-to-process";
 import { normalize } from "~/core/recognition/preprocessing";
 import { toSamples } from "~/core/recognition/recognizer/samples";
